+++ src/modules/crashrecovery/simplyquit/src/decrsqCoreFault.cpp
@@ -290,12 +290,14 @@ static void signalSegV( int number, siginfo_t *infos, void *ptrContext ){
 			char * const demangled = abi::__cxa_demangle( symbol, 0, 0, &status );
 			if( demangled ){
 				symbol.Format( "%s(%s+0x%x) [%p] %s", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr, demangled );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr, demangled );
 				free( demangled );
 				
 			}else{
 				symbol.Format( "%s(%s+0x%x) [%p]", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr );
 			}
 			
 		}else{
@@ -417,12 +419,14 @@ static void signalAbort( int number, siginfo_t *infos, void *ptrContext ){
 			char * const demangled = abi::__cxa_demangle( symbol, 0, 0, &status );
 			if( demangled ){
 				symbol.Format( "%s(%s+0x%x) [%p] %s", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr, demangled );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr, demangled );
 				free( demangled );
 				
 			}else{
 				symbol.Format( "%s(%s+0x%x) [%p]", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr );
 			}
 			
 		}else{
@@ -524,12 +528,14 @@ static void signalBusError( int number, siginfo_t *infos, void *ptrContext ){
 			char * const demangled = abi::__cxa_demangle( symbol, 0, 0, &status );
 			if( demangled ){
 				symbol.Format( "%s(%s+0x%x) [%p] %s", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr, demangled );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr, demangled );
 				free( demangled );
 				
 			}else{
 				symbol.Format( "%s(%s+0x%x) [%p]", info.dli_fname, info.dli_sname,
-					( const char* )addr - ( const char* )info.dli_saddr, addr );
+					( unsigned int )( ( const char* )addr - ( const char* )info.dli_saddr ),
+					addr );
 			}
 			
 		}else{
